IMPLEMENTED BLUEPRINTSgit status
